feat: チャンネルURL(/[channel])でアクセスできるようにし、ロゴクリックでホームへ戻す
- チャンネル視聴ページを /[channel] に分離し、URLのチャンネル名を起点にIRC接続を開始する
  (直接アクセス・検索フォームからの遷移の両方に対応)
- ホーム(/)はウェルカム画面専用にし、視聴中にロゴ(chat-sensei)クリックで戻ったら切断する
- チャンネル検索フォームは connect の直接呼び出しをやめ、正規化したチャンネル名の
  ページへ router.push で遷移する(接続はチャンネルページがURLを起点に行う)
- 切断(Disconnect)でストアのチャンネルが null に戻ったらホームへ遷移する
